refactor: simplify admin metrics code after review
- Extract relativeTime() to shared lib/utils.ts (was duplicated)
- Replace hardcoded PLAN_LIMITS with server-provided binLimit/storageLimit
- Static import for getPlanBreakdown (was unnecessary dynamic import)
- Cap lastActiveCache at 10k entries to prevent unbounded growth
- Parallelize independent queries in getPlanBreakdown with Promise.all
