+++ src/features/admin/AdminUsersTable.tsx
@@ -3,31 +3,9 @@ import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
 import { type SortDirection, SortHeader } from '@/components/ui/sort-header';
 import { Table, TableHeader, TableRow } from '@/components/ui/table';
-import { cn } from '@/lib/utils';
+import { cn, relativeTime } from '@/lib/utils';
 import type { AdminFieldKey } from './useAdminColumnVisibility';
 import { type AdminUser, capitalize, statusVariant } from './useAdminUsers';
-
-
-const PLAN_LIMITS: Record<string, { maxBins: number | null; maxStorageMb: number | null }> = {
-  free:  { maxBins: 50,   maxStorageMb: 0    },
-  plus:  { maxBins: 500,  maxStorageMb: 100  },
-  pro:   { maxBins: 5000, maxStorageMb: 1000 },
-};
-
-function relativeTime(iso: string | null): string {
-  if (!iso) return '—';
-  const ms = Date.now() - new Date(iso).getTime();
-  const mins = Math.floor(ms / 60000);
-  if (mins < 1) return 'just now';
-  if (mins < 60) return `${mins}m ago`;
-  const hours = Math.floor(mins / 60);
-  if (hours < 24) return `${hours}h ago`;
-  const days = Math.floor(hours / 24);
-  if (days < 30) return `${days}d ago`;
-  const months = Math.floor(days / 30);
-  if (months < 12) return `${months}mo ago`;
-  return `${Math.floor(months / 12)}y ago`;
-}
 
 function pctColor(pct: number): string {
   if (pct > 90) return 'text-[var(--destructive)]';
@@ -158,12 +136,11 @@ export function AdminUsersTable({
 
       {users.map((u) => {
         const isSelf = u.id === currentUserId;
-        const limits = PLAN_LIMITS[u.plan] ?? PLAN_LIMITS.free;
-        const binPct = limits.maxBins !== null && limits.maxBins > 0
-          ? Math.round((u.binCount / limits.maxBins) * 100)
+        const binPct = u.binLimit !== null && u.binLimit > 0
+          ? Math.round((u.binCount / u.binLimit) * 100)
           : null;
-        const storagePct = limits.maxStorageMb !== null && limits.maxStorageMb > 0
-          ? Math.round((u.photoStorageMb / limits.maxStorageMb) * 100)
+        const storagePct = u.storageLimit !== null && u.storageLimit > 0
+          ? Math.round((u.photoStorageMb / u.storageLimit) * 100)
           : null;
         const accountAge = Math.floor((Date.now() - new Date(u.createdAt).getTime()) / 86400000);
